Control diffuser to produce scent and colours
This is the first working version.  It provides a single synchronous call to
choose the colour emitted when a scent is released.  The range of colours that
can be produced is rather limited because each component (red, green and blue)
is controlled by a binary flag, even if the resulting signal is actually
encoded with 24 bits.  The effects of some parts of the signal are still
unknown to me, although I suspect that among other things they determine the
duration of the puff.