Copy signal data to audio stream with 'memcpy'
Initial tests on Linux fail as the audio signal drops, probably because
data is not processed quickly enough.  This new implementation of the
PortAudio callback that writes the signal to the output audio stream
replaces a 'for' loop with 'memcpy' and 'memset' calls.  This solution
should be faster as the end of the signal is not checked before every
frame copy.

On my Linux test environment, the generated audio signal has fewer drops
with this change.  I attempted to profile this optimization against the
previous commit with gprof and callgrind.  But the results are frankly
inconclusive, in part I believe because the overhead caused by those
tools leads to drops in the audio signal.  Finally, I performed
measurements with the standard 'clock' function.  Results were positive
but not particularly marked.