Encode consecutive colours in a signal
The existing implementation started playing the audio signal
asynchronously while the main thread waited 1 second and then closed the
stream.  This arbitrary long pause assumed that the audio signal sent to
the diffuser was shorter.

This change replaces the 'Pa_Sleep' call with a function that is called
by PortAudio once the signal has been entirely played.  This callback
uses a C++11 'future' object to notify the main thread that it can
resume execution.

But the Scentee diffuser does not recognize signals sent back-to-back in
the audio stream.  Nevertheless, code using the 'stinkee::Diffuser'
interface may call 'process' again, immediately after it returned.  To
address this issue, a 1 second silence gap is added after each signal.
Although sending a single colour to the diffuser is as slow as before,
it is now possible to compose a sequence of colours in a single signal
object.