Convert 'stinkee::Signal' class to a free function
There was little advantage to model the signal as an object, since it
did not maintain any state.  It was an unnecessary abstraction.

In addition, I prefer to give the application full access to the main
data structure.  Thus the user of the library could reserve a large
amount of memory up front for the signal or maybe reuse the same object
several times.